Add RequestConfirmationMessageSenders folder to \Areas\Admin and move message senders in there
- update StartUp to point to new location for message senders
