+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Notifications/RequestConfirmationsSentHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
+using AllReady.Areas.Admin.RequestConfirmationMessageSenders;
 using AllReady.Extensions;
 using AllReady.Models;
 using Hangfire;
